+++ packages/record/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @ohmymcp/record
+# @ohmymcp-hsu/record
 
 ## 0.1.1
 
@@ -19,11 +19,11 @@
 ### Patch Changes
 
 - Updated dependencies [0d92470]
-  - @ohmymcp/core@0.2.0
+  - @ohmymcp-hsu/core@0.2.0
 
 ## 0.0.1
 
 ### Patch Changes
 
 - Updated dependencies [606600f]
-  - @ohmymcp/core@0.1.0
+  - @ohmymcp-hsu/core@0.1.0